Granted scaladoc its own Global.
An incremental step on the road to disentangling
scaladoc from the compiler sources. This pushes the
elements in typer outward so scaladoc can subclass
Global like everyone else.
